Split Table Controllers, Update js  to point to controllers.
